move funciton::PyIterator -> protocol::PyIterIter
