+++ vm/src/protocol/mod.rs
@@ -3,5 +3,5 @@ mod iter;
 mod mapping;
 
 pub use buffer::{BufferInternal, BufferOptions, PyBuffer, ResizeGuard};
-pub use iter::{PyIter, PyIterReturn};
+pub use iter::{PyIter, PyIterIter, PyIterReturn};
 pub use mapping::{PyMapping, PyMappingMethods};
